Make project colliding with entity not cancellable
