+++ src/main/java/net/minestom/server/event/entity/projectile/ProjectileCollideEvent.java
@@ -2,16 +2,14 @@ package net.minestom.server.event.entity.projectile;
 
 import net.minestom.server.coordinate.Pos;
 import net.minestom.server.entity.Entity;
-import net.minestom.server.event.trait.CancellableEvent;
 import net.minestom.server.event.trait.EntityInstanceEvent;
 import net.minestom.server.event.trait.RecursiveEvent;
 import org.jetbrains.annotations.NotNull;
 
-class ProjectileCollideEvent implements EntityInstanceEvent, CancellableEvent, RecursiveEvent {
+class ProjectileCollideEvent implements EntityInstanceEvent, RecursiveEvent {
 
     private final @NotNull Entity projectile;
     private final @NotNull Pos position;
-    private boolean cancelled;
 
     protected ProjectileCollideEvent(@NotNull Entity projectile, @NotNull Pos position) {
         this.projectile = projectile;
@@ -25,15 +23,5 @@ class ProjectileCollideEvent implements EntityInstanceEvent, CancellableEvent, R
 
     public @NotNull Pos getCollisionPosition() {
         return position;
-    }
-
-    @Override
-    public boolean isCancelled() {
-        return cancelled;
-    }
-
-    @Override
-    public void setCancelled(boolean cancel) {
-        cancelled = cancel;
     }
 }
