+++ src/main/java/net/minestom/server/event/entity/projectile/ProjectileCollideWithBlockEvent.java
@@ -2,12 +2,14 @@ package net.minestom.server.event.entity.projectile;
 
 import net.minestom.server.coordinate.Pos;
 import net.minestom.server.entity.Entity;
+import net.minestom.server.event.trait.CancellableEvent;
 import net.minestom.server.instance.block.Block;
 import org.jetbrains.annotations.NotNull;
 
-public final class ProjectileCollideWithBlockEvent extends ProjectileCollideEvent {
+public final class ProjectileCollideWithBlockEvent extends ProjectileCollideEvent implements CancellableEvent {
 
     private final @NotNull Block block;
+    private boolean cancelled;
 
     public ProjectileCollideWithBlockEvent(
             @NotNull Entity projectile,
@@ -20,5 +22,15 @@ public final class ProjectileCollideWithBlockEvent extends ProjectileCollideEven
 
     public @NotNull Block getBlock() {
         return block;
+    }
+
+    @Override
+    public boolean isCancelled() {
+        return cancelled;
+    }
+
+    @Override
+    public void setCancelled(boolean cancel) {
+        cancelled = cancel;
     }
 }
